requirements.txt
fastapi==0.115.0
httpx==0.27.2
pydantic==2.9.2
psycopg[binary]==3.2.13
PyYAML==6.0.3
starlette==0.38.6
uvicorn==0.30.6
